+++ app-backend/migrations/src/main/scala/migrations/102.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/102.scala