+++ app-backend/migrations/src/main/scala/migrations/103.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/103.scala