+++ app-backend/migrations/src/main/scala/migrations/121.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/121.scala