+++ app-backend/migrations/src/main/scala/migrations/128.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/128.scala