+++ app-backend/migrations/src/main/scala/migrations/141.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/141.scala