+++ app-backend/migrations/src/main/scala/migrations/20.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/20.scala