+++ app-backend/migrations/src/main/scala/migrations/24.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/24.scala