+++ app-backend/migrations/src/main/scala/migrations/28.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/28.scala