+++ app-backend/migrations/src/main/scala/migrations/39.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/39.scala