+++ app-backend/migrations/src/main/scala/migrations/40.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/40.scala