+++ app-backend/migrations/src/main/scala/migrations/6.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/6.scala