+++ app-backend/migrations/src/main/scala/migrations/71.scala
@@ -1,1 +1,0 @@
-../../../../src_migrations/main/scala/71.scala