+++ app-backend/migrations/src_migrations/main/scala/140.scala
@@ -1,9 +1,0 @@
-import slick.jdbc.PostgresProfile.api._
-import com.liyaos.forklift.slick.SqlMigration
-
-object M140 {
-  RFMigrations.migrations = RFMigrations.migrations :+ SqlMigration(140)(
-    List(
-      sqlu"CREATE INDEX IF NOT EXISTS tile_footprint_idx ON scenes USING GIST (tile_footprint);"
-    ))
-}