+++ app-backend/migrations/src_migrations/main/scala/170.scala
@@ -1,9 +1,0 @@
-import slick.jdbc.PostgresProfile.api._
-import com.liyaos.forklift.slick.SqlMigration
-
-object M170 {
-  RFMigrations.migrations = RFMigrations.migrations :+ SqlMigration(170)(
-    List(
-      sqlu"DROP TABLE scenes_to_projects;"
-    ))
-}