+++ app-backend/migrations/src_migrations/main/scala/75.scala
@@ -1,9 +1,0 @@
-import slick.jdbc.PostgresProfile.api._
-import com.liyaos.forklift.slick.SqlMigration
-
-object M75 {
-  RFMigrations.migrations = RFMigrations.migrations :+ SqlMigration(75)(
-    List(
-      sqlu"ALTER TABLE uploads ADD COLUMN source text;"
-    ))
-}